Version 2.8.0
git-svn-id: https://plugins.svn.wordpress.org/super-socializer/trunk@920065 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ uninstall.php
@@ -3,28 +3,31 @@
 if( !defined( 'WP_UNINSTALL_PLUGIN' ) ) {
 	exit();
 }
-global $wpdb;
-$theChampOptions = array(
-					'the_champ_login',
-					'the_champ_facebook',
-					'the_champ_sharing',
-					'the_champ_ss_version'
-				);
-// For Single site
-if( !is_multisite() ){
-	foreach($theChampOptions as $option){
-		delete_option( $option );
-	}
-	$wpdb->query("delete from $wpdb->usermeta where meta_key like 'thechamp%'");
-}else{
-	// For Multisite
-	$theChampBlogIds = $wpdb->get_col( "SELECT blog_id FROM $wpdb->blogs" );
-	$theChampOriginalBlogId = get_current_blog_id();
-	foreach ( $theChampBlogIds as $blogId ){
-		switch_to_blog( $blogId );
+$theChampSharingOptions = get_option('the_champ_sharing');
+if(isset($theChampSharingOptions['delete_options'])){
+	global $wpdb;
+	$theChampOptions = array(
+						'the_champ_login',
+						'the_champ_facebook',
+						'the_champ_sharing',
+						'the_champ_ss_version'
+					);
+	// For Single site
+	if( !is_multisite() ){
 		foreach($theChampOptions as $option){
-			delete_site_option($option);
+			delete_option( $option );
 		}
+		$wpdb->query("delete from $wpdb->usermeta where meta_key like 'thechamp%'");
+	}else{
+		// For Multisite
+		$theChampBlogIds = $wpdb->get_col( "SELECT blog_id FROM $wpdb->blogs" );
+		$theChampOriginalBlogId = get_current_blog_id();
+		foreach ( $theChampBlogIds as $blogId ){
+			switch_to_blog( $blogId );
+			foreach($theChampOptions as $option){
+				delete_site_option($option);
+			}
+		}
+		switch_to_blog( $theChampOriginalBlogId );
 	}
-	switch_to_blog( $theChampOriginalBlogId );
 }
